fix(6313): code cleanup and debug edge cases with use of durations without negative time
