+++ src/languageSupport/languages/agnostic/types.ts
@@ -1,0 +1,9 @@
+export interface LspRange {
+  start: LspPosition
+  end: LspPosition
+}
+
+export interface LspPosition {
+  column: number
+  line: number
+}
